fix: cast provided_quantity to float8 for accurate max_trf_quantity calculation

+++ src/routes/other/inventory/purchase_entry/handlers.ts
@@ -16,7 +16,7 @@ export const valueLabel: AppRouteHandler<ValueLabelRoute> = async (c: any) => {
     .select({
       value: purchase_entry.uuid,
       label: sql`CONCAT( ${product.name}, ' - ', ${purchase_entry.serial_no})`,
-      max_trf_quantity: sql`${purchase_entry.quantity} - COALESCE(${purchase_return_entry.quantity}, 0) - COALESCE(${purchase_entry.provided_quantity}, 0)`,
+      max_trf_quantity: sql`${purchase_entry.quantity} - COALESCE(${purchase_return_entry.quantity}, 0) - COALESCE(${purchase_entry.provided_quantity}, 0)::float8`,
       // warehouse_uuid: purchase_entry.warehouse_uuid,
       // warehouse_name: warehouse.name,
     })
